% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomarker_outliers.R
\name{calculate_outlier_bounds}
\alias{calculate_outlier_bounds}
\title{Biomarker Outlier Functions}
\usage{
calculate_outlier_bounds(dataf, log_10_biom_column)
}
\arguments{
\item{dataf}{A data frame}

\item{log_10_biom_column}{Character. Name of the log10-transformed biomarker column}
}
\value{
Data frame with lower bound (LB), upper bound (UB), and notes
}
\description{
Functions for identifying and handling outliers in biomarker data.
Calculates outlier bounds for biomarkers
}
\details{
Calculates lower and upper bounds for outlier detection using the
IQR method (Q1 - 1.5\emph{IQR and Q3 + 1.5}IQR).
}
