% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apoe_validation.R
\name{.add_e4_validation_columns}
\alias{.add_e4_validation_columns}
\title{Add e4 validation columns}
\usage{
.add_e4_validation_columns(dataf, genotype_col, e4_col)
}
\description{
Add e4 validation columns
}
\keyword{internal}
