% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apoe_validation.R
\name{.add_snp_validation_columns}
\alias{.add_snp_validation_columns}
\title{Add SNP validation columns}
\usage{
.add_snp_validation_columns(dataf, rs7412_col, rs429358_col, genotype_col)
}
\description{
Add SNP validation columns
}
\keyword{internal}
