% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{.APOE_RISK_GROUPS}
\alias{.APOE_RISK_GROUPS}
\title{APOE risk group levels (in order for factor levels and reporting)}
\format{
An object of class \code{character} of length 3.
}
\usage{
.APOE_RISK_GROUPS
}
\description{
APOE risk group levels (in order for factor levels and reporting)
}
\keyword{internal}
