% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apoe_genetics_helpers.R
\name{.classify_apoe_to_risk_group}
\alias{.classify_apoe_to_risk_group}
\title{Classify APOE genotype to risk group}
\usage{
.classify_apoe_to_risk_group(std_genotypes)
}
\arguments{
\item{std_genotypes}{Vector of standardized APOE genotypes}
}
\value{
Vector of risk group classifications
}
\description{
Classify APOE genotype to risk group
}
\keyword{internal}
