% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apoe_genetics_helpers.R
\name{.clean_snp_genotype_values}
\alias{.clean_snp_genotype_values}
\title{Clean SNP genotype values to handle CT/TC equivalency}
\usage{
.clean_snp_genotype_values(snp_vector)
}
\arguments{
\item{snp_vector}{Vector of SNP values}
}
\value{
Cleaned SNP values with TC converted to CT
}
\description{
Clean SNP genotype values to handle CT/TC equivalency
}
\keyword{internal}
