% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitive_composites.R
\name{.compute_cognitive_composites}
\alias{.compute_cognitive_composites}
\title{Compute composite scores from z-scores}
\usage{
.compute_cognitive_composites(dataf, test_groups)
}
\description{
Compute composite scores from z-scores
}
\keyword{internal}
