% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitive_composites.R
\name{.compute_cognitive_zscores}
\alias{.compute_cognitive_zscores}
\title{Compute z-scores grouped by demographic variables for cognitive variables}
\usage{
.compute_cognitive_zscores(dataf, test_groups, grouping_vars)
}
\description{
Compute z-scores grouped by demographic variables for cognitive variables
}
\keyword{internal}
