% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitive_composites.R
\name{.compute_composite_scores}
\alias{.compute_composite_scores}
\title{Compute composite scores from z-scores}
\usage{
.compute_composite_scores(dataf, test_groups)
}
\description{
Compute composite scores from z-scores
}
\keyword{internal}
