% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitive_composites.R, R/z_scores.R
\name{.compute_grouped_zscores}
\alias{.compute_grouped_zscores}
\title{Compute z-scores grouped by demographic variables}
\usage{
.compute_grouped_zscores(dataf, vars, group_vars, prefix)

.compute_grouped_zscores(dataf, vars, group_vars, prefix)
}
\description{
Compute z-scores grouped by demographic variables

Compute grouped z-scores
}
\keyword{internal}
