% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_scores.R
\name{.compute_simple_zscores}
\alias{.compute_simple_zscores}
\title{Compute simple z-scores (no grouping)}
\usage{
.compute_simple_zscores(dataf, vars, prefix)
}
\description{
Compute simple z-scores (no grouping)
}
\keyword{internal}
