% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apoe_genetics_helpers.R
\name{.create_snp_to_apoe_lookup}
\alias{.create_snp_to_apoe_lookup}
\title{Create SNP to APOE genotype lookup table}
\usage{
.create_snp_to_apoe_lookup()
}
\value{
Data frame with SNP combinations and predicted APOE genotypes
}
\description{
Create SNP to APOE genotype lookup table
}
\keyword{internal}
