% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{.E4_CARRIER_GENOTYPES}
\alias{.E4_CARRIER_GENOTYPES}
\title{APOE genotype carrier groups}
\format{
An object of class \code{character} of length 3.
}
\usage{
.E4_CARRIER_GENOTYPES
}
\description{
APOE genotype carrier groups
}
\keyword{internal}
