% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apoe_validation.R
\name{.format_apoe_genotype}
\alias{.format_apoe_genotype}
\title{Format APOE genotype to standard format (E2E3, E3E4, etc.)}
\usage{
.format_apoe_genotype(genotypes)
}
\description{
Format APOE genotype to standard format (E2E3, E3E4, etc.)
}
\keyword{internal}
