% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinical_calculations.R
\name{.get_expected_bp_outputs}
\alias{.get_expected_bp_outputs}
\title{Determine expected BP output columns based on user options}
\usage{
.get_expected_bp_outputs(calculate_map, calculate_pp)
}
\description{
Determine expected BP output columns based on user options
}
\keyword{internal}
