% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitive_composites.R
\name{.get_expected_composite_outputs}
\alias{.get_expected_composite_outputs}
\title{Get expected output columns for composite scores}
\usage{
.get_expected_composite_outputs(test_groups)
}
\description{
Get expected output columns for composite scores
}
\keyword{internal}
