% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apoe_genetics_helpers.R
\name{.predict_apoe_from_snps}
\alias{.predict_apoe_from_snps}
\title{Predict APOE genotype from SNP values}
\usage{
.predict_apoe_from_snps(rs7412_values, rs429358_values, lookup_table)
}
\arguments{
\item{rs7412_values}{Vector of rs7412 values}

\item{rs429358_values}{Vector of rs429358 values}

\item{lookup_table}{Lookup table for SNP combinations}
}
\value{
Vector of predicted APOE genotypes
}
\description{
Predict APOE genotype from SNP values
}
\keyword{internal}
