% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apoe_validation.R
\name{.predict_genotype_from_snps}
\alias{.predict_genotype_from_snps}
\title{Predict APOE genotype from SNP combinations}
\usage{
.predict_genotype_from_snps(rs7412, rs429358)
}
\description{
Predict APOE genotype from SNP combinations
}
\keyword{internal}
