% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.R
\name{.replace_outliers_single_var}
\alias{.replace_outliers_single_var}
\title{Replace outliers for a single variable}
\usage{
.replace_outliers_single_var(dataf, var, multiplier, paired_cols)
}
\description{
Replace outliers for a single variable
}
\keyword{internal}
