% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apoe_validation.R
\name{.report_risk_classification}
\alias{.report_risk_classification}
\title{Report risk group classification}
\usage{
.report_risk_classification(result_df, group_col)
}
\description{
Report risk group classification
}
\keyword{internal}
