% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apoe_validation.R
\name{.return_validation_results}
\alias{.return_validation_results}
\title{Return validation results with consistent messaging}
\usage{
.return_validation_results(
  result_df,
  original_df,
  validation_col,
  return_all,
  validation_type
)
}
\description{
Return validation results with consistent messaging
}
\keyword{internal}
