% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{.SNP_GENOTYPE_MAP}
\alias{.SNP_GENOTYPE_MAP}
\title{SNP to APOE genotype mapping
Maps rs7412 and rs429358 SNP combinations to APOE genotypes}
\format{
An object of class \code{list} of length 9.
}
\usage{
.SNP_GENOTYPE_MAP
}
\description{
SNP to APOE genotype mapping
Maps rs7412 and rs429358 SNP combinations to APOE genotypes
}
\keyword{internal}
