% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/med_rxcui.R
\name{find_approx_rxcui}
\alias{find_approx_rxcui}
\title{Direct API call to RxNorm for approximate RxCUI lookup}
\usage{
find_approx_rxcui(term, maxEntries = 3)
}
\arguments{
\item{term}{Character. Medication term to look up}

\item{maxEntries}{Numeric. Maximum number of entries to return (default: 3)}
}
\value{
Character. RxCUI if found, NA_character_ otherwise
}
\description{
Direct API call to RxNorm for approximate RxCUI lookup
}
