% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medication_rxcui_helpers.R
\name{null_coalesce}
\alias{null_coalesce}
\title{Null-coalescing operator}
\usage{
null_coalesce(x, y)
}
\arguments{
\item{x}{Any value to test}

\item{y}{Default value to return if x is NULL or NA}
}
\value{
x if not NULL/NA, otherwise y
}
\description{
Returns the right-hand side value if the left-hand side is NULL or NA,
otherwise returns the left-hand side value. Similar to \code{||} in other languages.
}
\keyword{internal}
