% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medication_atc2_helpers.R
\name{add_medication_atc2_nested}
\alias{add_medication_atc2_nested}
\title{Add ATC2 classifications to medications (nested version)}
\usage{
add_medication_atc2_nested(
  dataf,
  rxcui_col = "rxcui",
  new_col_name = "atc2_class",
  ...
)
}
\description{
Add ATC2 classifications to medications (nested version)
}
\keyword{internal}
