% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medication_atc2_helpers.R
\name{add_medication_atc2_unnested}
\alias{add_medication_atc2_unnested}
\title{Add ATC2 classifications to medications (unnested version)}
\usage{
add_medication_atc2_unnested(
  dataf,
  rxcui_col = "rxcui",
  new_col_name = "atc2_class",
  ...
)
}
\description{
Add ATC2 classifications to medications (unnested version)
}
\keyword{internal}
