% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apoe_genetics.R
\name{classify_apoe_risk_groups}
\alias{classify_apoe_risk_groups}
\title{Classify APOE genotypes into risk groups}
\usage{
classify_apoe_risk_groups(dataf, genotype_col, group_col = "apoe_risk_group")
}
\arguments{
\item{dataf}{A data frame containing APOE genotype data}

\item{genotype_col}{Character. Name of the genotype column}

\item{group_col}{Character. Name for new risk group column (default: "apoe_risk_group")}
}
\value{
Data frame with added risk group column
}
\description{
Classify APOE genotypes into risk groups
}
