% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_scores.R
\name{.compute_grouped_zscores}
\alias{.compute_grouped_zscores}
\title{Compute grouped z-scores}
\usage{
.compute_grouped_zscores(dataf, vars, group_vars, prefix)
}
\description{
Compute grouped z-scores
}
\keyword{internal}
