% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apoe_genetics_helpers.R
\name{.format_apoe_genotype}
\alias{.format_apoe_genotype}
\title{Format APOE genotype to standard E2E3 notation}
\usage{
.format_apoe_genotype(
  genotype_vector,
  format = c("standard", "numeric", "slash")
)
}
\arguments{
\item{genotype_vector}{Vector of genotype values to format}

\item{format}{Format of the input genotypes}
}
\value{
Vector of formatted genotypes in E2E3 format
}
\description{
Format APOE genotype to standard E2E3 notation
}
\keyword{internal}
