% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apoe_genetics.R
\name{match_snp_genotype}
\alias{match_snp_genotype}
\title{Match APOE SNPs to genotypes}
\usage{
match_snp_genotype(
  dataf,
  rs7412_col,
  rs429358_col,
  genotype_col,
  return_all = TRUE
)
}
\arguments{
\item{dataf}{A data frame containing SNPs and genotypes}

\item{rs7412_col}{Character. Name of the rs7412 SNP column}

\item{rs429358_col}{Character. Name of the rs429358 SNP column}

\item{genotype_col}{Character. Name of the genotype column}

\item{return_all}{Logical. Return entire dataset or just mismatches (default: TRUE)}
}
\value{
Data frame with validation results
}
\description{
Match APOE SNPs to genotypes
}
